text input for title modifies ember data. date input is proving difficult
I'm still not sure how the browser interacts with the date input data. It looks it sends a this.set() when you first click it, and another one when you finish entering the data. this would be alright if it wasn't for the fact that I'm setting the retrieved data directly to dueDate. I think the best course of action now would be to create an interim object where I can do input validation before pushing that object to the datastore. If not, the least I can do is to save the date and time to a new object then combine the two into a string and pass it to dueDate from there.